Update examples and code for bool API changes

+++ src/connection.rs
@@ -645,7 +645,7 @@ impl ExtensionInformation {
                     let info = info.and_then(|c| c.reply().ok());
                     if let Some(info) = info {
                         // If the extension is not present, we return None, else we box it
-                        if info.present == 0 {
+                        if info.present as u8 == 0 {
                             None
                         } else {
                             Some(info)
